python-backend/requirements.txt
pymongo
langchain
langchain-community
langchain-text-splitters
psycopg2-binary
pgvector
grpcio
grpcio-tools
sentence-transformers